refactor: Include location aliases in manuscript view and toponym view
This commit modifies the manuscript view and toponym view to include location aliases. In the manuscript view, the related LocationAlias objects are fetched for each location mentioned in the folios and added to the location object. In the toponym view, the unique and sorted LocationAlias objects based on placename_modern are retrieved and passed to the template as "aliases". This change enhances the display of location information in both views.